Import wiggle data generated by the lab's ChIP-seq analysis pipeline

Given a path to a collection of 16 tab-separated wiggle data files (one per yeast chromosome), imports data into a GRanges object.

import_wiggle(path, local_copy = TRUE)

Arguments

path
Path to the folder containing the wiggle data. No default.

local_copy
Logical indicating whether to create local copy of target file before importing data. If TRUE a local copy is automatically created and deleted after use. Use this argument to avoid reading files directly from shared locations (namely LabShare). Defaults to TRUE.

Value

GRanges object.

Examples

# NOT RUN {
x <- import_wiggle('wiggle_data/')

dot1 <- import_wiggle('/Volumes/LabShare/HTGenomics/HiSeqOutputs/\
                      AveReps_SacCer3_MACS2/Red1-dot1D-195-16-reps-MACS2/',
                      local_copy=TRUE)
# }