Signal around provided genomic positions

Pulls out signal in intervals of the defined length centered at provided positions. Note: Our ChIP-seq data always contains gaps with missing data. The affected intervals will contain "NA" values in the output. As a warning, the number of affected genes is printed to the console.

signal_at_position(signal_data, positions, position_names, up_ext = 1000,
  down_ext = 1000, write_to_file = FALSE)

Arguments

signal_data
Input signal track data as a GRanges object (see ?"GRanges-class" for more details). To load wiggle and bedGraph data run import_wiggle and import_bedGraph, respectively. No default.

positions
Position data as a GRanges object. Will also accept an interval GRanges object of constant width (extended by specified up_ext and down_ext). No default.

position_names
String indicating the column containing the position names: either the GRanges 'seqnames' or a metadata column. Will be added to the output to identify the different positions. No default.

up_ext
Integer indicating number of bp to collect upstream of start position. Defaults to 1000.

down_ext
Integer indicating number of bp to collect downstream of end position. Deafaults to 1000.

write_to_file
Logical indicating whether output should be written to a .txt file (in current working directory). If to_file = FALSE the function returns output R object. Defaults to FALSE.

Value

A data frame with 1001 columns:

gene First column with the gene IDs (if found in input gff)

t1 to t1000 1000 columns with signal at each meta position.

Columns t1 to t1000 correspond to the following meta positions:

t1 to t250 Upstream flank (1/2 of gene length)

t251 to t750 ORF

t751 to t1000 Downstream flank (1/2 of gene length)

Examples

not_run({
  signal_at_position(WT, Red1_summits, 'name')

  signal_at_position(WT, centromeres, 'seqnames',
                     up_ext = 5000, down_ext = 5000,
                     write_to_file = TRUE)
})