Signal flanking rDNA

Pulls out ChIP signal at specified window sizes flanking the rDNA region on chromosome 12.

signal_flanking_rDNA(signal_data, flank_length = 40000, genome)

Arguments

signal_data
Input signal track data as a GRanges object (see ?"GRanges-class" for more details). To load wiggle and bedGraph data run import_wiggle and import_bedGraph, respectively. No default.

flank_length
Integer specifying the length (in bp) of the windows to collect signal for up and downstream of the rDNA. Defaults to 40000 (i.e. 40 kb).

genome
Character object specifying the genome version; accepts one of the following options:

"SK1Yue"

"sacCer3"

"SK1"

No default.

Examples

not_run({
  signal_flanking_rDNA(WT, genome = 'SK1Yue')

  signal_flanking_rDNA(WT, flank_length=50000, genome = 'sacCer3')
})